+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 All notable changes to this project will be documented here.
 
 ## [Unreleased]
+### Added
+- Memory-mapping made optional.
 
 ## [0.7.2] - 2020-03-21
 ### Added